requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
werkzeug==2.3.7
apscheduler==3.10.4
africastalking==2.0.2
gunicorn==21.2.0
